prediction_service/requirements.txt
fastapi
pydantic
grpcio
joblib
confluent-kafka
uvicorn
scikit-learn
joblib  
numpy
pandas
protobuf
grpcio-tools
prometheus-client
